requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
python-calamine
pytest
httpx
python-dotenv
email-validator
python-dateutil
